I realized we were making _a lot_ (hundreds) of **isMaster** and **serverStatus** calls to the server per spec test. Profiler results for transaction tests (see images below) show that currently, 16.70% of the time is spent in the function `should_skip_spec_tests`. That function makes multiple roundtrips to the server. After this change, only 2.26% of the time is spent in `should_skip_spec_tests`—a pretty big improvement. 

I also timed `test_transactions_spec` locally before and after caching server results. Here's the results:

```
Before optimization
n=10 | min    | Q1     | med    | mean   | Q3     | max    | std
wall | 28.30  | 28.44  | 28.63  | 28.70  | 28.88  | 29.60  | 0.3865
user | 0.6500 | 0.6550 | 0.6750 | 0.6770 | 0.6975 | 0.7100 | 0.0226
sys  | 0.110  | 0.120  | 0.140  | 0.138  | 0.150  | 0.170  | 0.0230
rss  | 13368  | 13408  | 13556  | 13564  | 13722  | 13740  | 159.2700

After optimization
n=10 | min   | Q1    | med   | mean  | Q3    | max   | std
wall | 27.66 | 27.70 | 27.89 | 28.03 | 28.13 | 28.84 | 0.4206
user | 0.500 | 0.520 | 0.520 | 0.524 | 0.530 | 0.550 | 0.0143
sys  | 0.080 | 0.090 | 0.100 | 0.098 | 0.100 | 0.120 | 0.0140
rss  | 13544 | 13635 | 13740 | 13735 | 13860 | 13908 | 136.2227
```
The transaction tests were run 10 times (n=10) using a script which called `time test_transactions_spec` on Linux and used a Rscript to calculate the mean, median, etc.

Profile results from CLion:
![before-optimization](https://user-images.githubusercontent.com/34226620/108432285-22937900-7212-11eb-84a0-c59342e8c2a1.png)
![after-optimization](https://user-images.githubusercontent.com/34226620/108432290-245d3c80-7212-11eb-92ec-74236b1c8f15.png)