@kevinAlbs 2021-02-19
> After this change, only 2.26% of the time is spent in should_skip_spec_tests—a pretty big improvement.

That is awesome. CLion's flame graph profiler does look really useful for drilling down into perf.

I'm not sure how much this translates into total runtime saved. I looked at tasks in the Evergreen patch, which shows `Duration` and `Base commit duration`. Some of the Windows tasks looked like a possible significant improvement (e.g. [1 minute difference](https://spruce.mongodb.com/task/cxx_driver_windows_2k8_release_compile_and_test_with_shared_libs_extra_alignment_patch_3c7609f04488c547ad7e1607f2eb8b675a107e83_602eec3b3e8e861950994c00_21_02_18_22_37_48/logs?execution=0)). Some other tasks looked like a difference of +/- 30 seconds.

Regardless of the impact on test runtime, this seems like a positive change IMO (to centralize `isMaster` and `serverStatus` calls).

> The transaction tests were run 10 times (n=10) using a script which called time test_transactions_spec on Linux and used a Rscript to calculate the mean, median, etc.

Nice analysis!